+++ CHANGELOG.md
@@ -10,7 +10,10 @@ Version 0.1.1
 - created poetry environment
 - created django project : intelliesg
 - created django app : envdata
-- 
+- created db connection: postgresql
+- created the dotenv file
+- added whitenoise for staticfiles management
+- finished configuring settings.py 
 
 
 Version 0.1.0